+++ crates/services/txpool/src/mock_db.rs
@@ -1,5 +1,8 @@
 use crate::ports::TxPoolDb;
-use fuel_core_storage::Result as StorageResult;
+use fuel_core_storage::{
+    transactional::AtomicView,
+    Result as StorageResult,
+};
 use fuel_core_types::{
     entities::{
         coins::coin::{
@@ -91,8 +94,18 @@ impl TxPoolDb for MockDb {
     fn is_message_spent(&self, id: &Nonce) -> StorageResult<bool> {
         Ok(self.data.lock().unwrap().spent_messages.contains(id))
     }
+}
 
-    fn current_block_height(&self) -> StorageResult<BlockHeight> {
-        Ok(Default::default())
+pub struct MockDBProvider(pub MockDb);
+
+impl AtomicView for MockDBProvider {
+    type View = MockDb;
+
+    fn view_at(&self, _: BlockHeight) -> StorageResult<Self::View> {
+        Ok(self.latest_view())
+    }
+
+    fn latest_view(&self) -> Self::View {
+        self.0.clone()
     }
 }